dev: chg: refactored pages into parts
