+++ pages/_sidebar.py
@@ -1,28 +1,33 @@
 import streamlit as st
 from typing import cast
 
-# headline sidebar
-st.sidebar.write("## Settings :gear:")
-
-# decide which version and language
-version_options = {"simple": "Simple", "expert": "Expert"}
+# FIXME: this should probably be external data?
+VERSION_OPTIONS = {"simple": "Simple", "expert": "Expert"}
 
 
-version: str = cast(
-    str,
-    st.sidebar.selectbox(
-        "⚙ Version 👩‍🎓👩🏼‍🔬",
-        version_options.keys(),
-        format_func=lambda x: version_options[x],
-    ),
-)
+def make_sidebar() -> tuple[str, str]:
+    st.sidebar.write("## Settings :gear:")
 
-# language selectbox
-language: str = cast(
-    str,
-    st.sidebar.selectbox(
-        "⚙ Language 🌍💬",
-        ["English", "German"],
-        label_visibility="visible",
-    ),
-)
+    version: str = cast(
+        str,
+        st.sidebar.selectbox(
+            "⚙ Version 👩‍🎓👩🏼‍🔬",
+            VERSION_OPTIONS.keys(),
+            format_func=lambda x: VERSION_OPTIONS[x],
+        ),
+    )
+
+    # language selectbox
+    language: str = cast(
+        str,
+        st.sidebar.selectbox(
+            "⚙ Language 🌍💬",
+            ["English", "German"],
+            label_visibility="visible",
+        ),
+    )
+    return version, language
+
+
+if __name__ == "__main__":
+    make_sidebar()
